Move common files from uttaksplan to common
