+++ src/common/components/formatert-dato/formatertDato.less
@@ -1,4 +1,4 @@
-@import '../../styles/variabler.less';
+@import '../../styles/variables.less';
 
 .formatert-dato {
     &__ukedag {
