+++ src/common/components/range-input/rangeInput.less
@@ -1,4 +1,4 @@
-@import '../../styles/variabler.less';
+@import '../../styles/variables.less';
 
 .rangeInputStyling(@width: 100%, @trackColor: #66cbec, @thumbBkg: white, @thumbBorder: #0067c5) {
     @thumbSize: 32px;
