+++ src/common/components/range-input/rangeStepper.less
@@ -1,4 +1,5 @@
-@import '../../styles/variabler.less';
+@import '../../styles/variables.less';
+@import '../../styles/mixins.less';
 
 .rangeStepper {
     .borderButton();
