+++ src/common/components/sirkelknapp/sirkelknapp.less
@@ -1,4 +1,5 @@
-@import '../../styles/variabler.less';
+@import '../../styles/variables.less';
+@import '../../styles/mixins.less';
 
 .sirkelknappBla() {
     background-color: @c-buttonBlue;
