--- src/common/util/animation.ts
@@ -1,1 +1,0 @@
-export const collapseSpringConfig = { stiffness: 250, damping: 30 };
