+++ src/common/util/bem.ts
