+++ src/uttaksplan/components/timeline/items/MarkerItem.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
 import { TimelineMarker } from 'uttaksplan/components/timeline/types';
-import BEMHelper from 'uttaksplan/utils/bem';
 import TimelineIcons from 'uttaksplan/components/timeline/TimelineIcons';
-import Dato from 'uttaksplan/elements/dato/Dato';
 import { TimelineItemProps } from 'uttaksplan/components/timeline/Timeline';
+import FormatertDato from 'common/components/formatert-dato/FormatertDato';
+import BEMHelper from 'common/util/bem';
 
 export interface Props extends TimelineItemProps {
     item: TimelineMarker;
@@ -19,7 +19,7 @@ const MarkerItem: React.StatelessComponent<Props> = ({
         <TimelineIcons icons={item.icons} iconRenderer={iconRenderer} />
         <h1 className={BEM.element('title')}>{item.title}</h1>
         <div className={BEM.element('date')}>
-            <Dato dato={item.date} />
+            <FormatertDato dato={item.date} />
         </div>
         <div className={BEM.element('comment')}> {item.comment}</div>
     </article>
