+++ src/uttaksplan/components/utsettelseSkjema/UtsettelseSkjema.tsx
@@ -11,10 +11,9 @@ import {
     Tidsperiode,
     Permisjonsregler
 } from 'uttaksplan/types';
-import { normaliserDato, preventFormSubmit } from 'uttaksplan/utils';
+import { preventFormSubmit } from 'uttaksplan/utils';
 import { isBefore, isSameDay } from 'date-fns';
 import Ferieinfo from 'uttaksplan/components/utsettelseSkjema/Ferieinfo';
-import { renderDag } from 'uttaksplan/utils/renderUtils';
 import {
     validerUtsettelseskjema,
     getTilTidsromSluttdato,
@@ -33,6 +32,8 @@ import Radioliste from 'uttaksplan/components/radioliste/Radioliste';
 import EkspanderbartInnhold from 'uttaksplan/elements/ekspanderbartInnhold/EkspanderbartInnhold';
 import { SkjemaGruppe } from 'nav-frontend-skjema';
 import DatoInput from 'common/components/dato-input/DatoInput';
+import { renderDag } from 'common/util/renderUtils';
+import { normaliserDato } from 'common/util/datoUtils';
 
 interface OwnProps {
     termindato: Date;
