+++ src/uttaksplan/redux/reducers/viewReducer.ts
@@ -16,7 +16,7 @@ export enum Infotekster {
 
 const defaultState: UttaksplanViewState = {
     synligInfo: new Map(),
-    spraak: 'nb',
+    Språkkode: 'nb',
     visTidslinje: false
 };
 
@@ -44,8 +44,6 @@ const updateState = (
 
 const ViewReducer = (state = defaultState, action: PlanleggerActionTypes) => {
     switch (action.type) {
-        case PlanleggerActionTypeKeys.SET_SPRAAK:
-            return updateState(state, { spraak: action.spraak });
         case PlanleggerActionTypeKeys.INFO_VIS:
             return updateState(state, {
                 synligInfo: leggTilInfo(state.synligInfo, action.id)
