--- src/uttaksplan/styles/_dev.less
@@ -1,33 +1,0 @@
-.intlCallout() {
-    &:after {
-        content: attr(data-key);
-        position: absolute;
-        z-index: 2001;
-        bottom: -1.3rem;
-        left: 0;
-        background: @c-purple;
-        color: white;
-        box-shadow: 0 0 0.5rem rgba(0, 0, 0, 0.2);
-        padding: 0.1rem 0.5rem;
-        border-radius: 2px;
-        font-size: 0.9rem;
-        font-weight: 200;
-        opacity: 0.9;
-        transition: all 0.2s ease;
-    }
-    &:hover {
-        &:after {
-            opacity: 1;
-            font-size: 1rem;
-            padding: 0.4rem 0.8rem;
-            bottom: -1.5rem;
-            left: -0.3rem;
-        }
-    }
-}
-
-body.visIntlKoder .intlTekst {
-    position: relative;
-    outline: 1px dotted @c-purple;
-    .intlCallout();
-}
