+++ src/uttaksplan/styles/planlegger.less
@@ -6,4 +6,3 @@
 @import './print.less';
 
 @import './app.less';
-@import './_dev.less';
