+++ src/uttaksplan/types/index.ts
@@ -1,17 +1,7 @@
 import { Periode } from './periodetyper';
 
 export * from './periodetyper';
-
-export type Dekningsgrad = '80%' | '100%';
-
-export type Forelder = 'forelder1' | 'forelder2';
-
-export type Spraak = 'nb' | 'nn';
-
-export interface Tidsperiode {
-    startdato: Date;
-    sluttdato: Date;
-}
+export * from '../../common/types';
 
 export interface Permisjonsregler {
     /** Totalt antall uker ved 80% */
